fasiliteter og funksjoner added

+++ Controllers/ajax/funksjoner_fasiliteter.php
@@ -50,9 +50,17 @@ if (isset($_POST['rem_funksjoner'])) {
     $frm_data = filteration($_POST);
     $values = [$frm_data['rem_funksjoner']];
 
+    $check_q = select('SELECT * FROM `rom_funksjoner` WHERE `funksjoner_id` = ?', [$frm_data['rem_funksjoner']], "i");
+
+   if (mysqli_num_rows($check_q)==0) {
+   
     $q = "DELETE FROM `funksjoner` WHERE `id` = ?";
     $res = delete($q, $values, "i");
     echo $res;
+   }else{
+       echo 'rom_added';
+
+   
 }
 
 
@@ -121,19 +129,31 @@ if (isset($_POST['rem_fasiliteter'])) {
     $frm_data = filteration($_POST);
     $values = [$frm_data['rem_fasiliteter']];
 
+     $check_q = select('SELECT * FROM `rom_fasiliteter` WHERE `fasiliteter_id` = ?', [$frm_data['rem_fasiliteter']], "i");
+     
+     if (mysqli_num_rows($check_q)==0) {
 
 
-    $pre_q = "SELECT * FROM `fasiliteter` WHERE `id` = ?";
-    $res = select($pre_q, $values, "i");
-    $img = mysqli_fetch_assoc($res);
-
-    if (deleteImage($img['icon'], FASILITETER_FOLDER)) {
-
-
-        $q = "DELETE FROM `fasiliteter` WHERE `id` = ?";
-        $res = delete($q, $values, "i");
-        echo $res;
-    } else {
-        echo 0;
+        $pre_q = "SELECT * FROM `fasiliteter` WHERE `id` = ?";
+        $res = select($pre_q, $values, "i");
+        $img = mysqli_fetch_assoc($res);
+    
+        if (deleteImage($img['icon'], FASILITETER_FOLDER)) {
+    
+    
+            $q = "DELETE FROM `fasiliteter` WHERE `id` = ?";
+            $res = delete($q, $values, "i");
+            echo $res;
+        } else {
+            echo 0;
+        }
     }
+
+        
+     } else {
+        echo 'rom_added';
+
+  
+     }
 }
+ ?>
